♻️ refactor: refactor service directory
